+++ tests/predis/standalone/data/ar/CustomerQuery.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace yiiunit\extensions\redis\predis\standalone\data\ar;
+
+use yii\redis\ActiveQuery;
+
+/**
+ * CustomerQuery
+ */
+class CustomerQuery extends ActiveQuery
+{
+    /**
+     * @return $this
+     */
+    public function active()
+    {
+        $this->andWhere(['status' => 1]);
+
+        return $this;
+    }
+}